requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
kombu==5.2.4
Pillow==9.1.1
prompt-toolkit==3.0.29
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
